src/requirements.txt
aiohttp==3.12.9
aioredis==2.0.1
alembic==1.16.3
asyncpg==0.30.0
click==8.2.1
cryptography==45.0.5
fastapi==0.116.0
httpx==0.28.1
motor==3.7.1
numpy==2.3.1
pandas==2.3.1
prometheus_client==0.22.1
psutil==7.0.0
psycopg2_binary==2.9.10
pydantic==2.11.7
PyJWT==2.10.1
pytest==8.4.0
python-dotenv==1.1.1
python_bcrypt==0.3.2
PyYAML==6.0.2
redis==6.2.0
Requests==2.32.4
scipy==1.16.0
SQLAlchemy==2.0.41
structlog==25.4.0
tenacity==9.1.2
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
